Move delete button to own file [PAR-88]
