+++ src/SettingsPage.tsx
@@ -1,20 +1,18 @@
 import React, { useState, useContext, useEffect, useCallback } from "react";
 import { useFormik, FormikErrors } from "formik";
-import styled from "styled-components";
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 import Col from "react-bootstrap/Col";
 import Collapse from "react-bootstrap/Collapse";
-import { HTMLInputElement } from "react-bootstrap";
 
 import PageLayout from "./PageLayout";
 import { AuthContext } from "./AuthContextProvider";
 import { User } from "./ourtypes";
 import { toUser } from "./mappers";
-import ConfirmationContext from "./ConfirmationContext";
+import DeleteAccountButton from "./DeleteAccountButton";
 
 const SettingsPage: React.FC = () => {
-  const { credential, logout } = useContext(AuthContext);
+  const { credential } = useContext(AuthContext);
   const [user, setUser] = useState<User>({ id: "", name: "", status: null });
   const [isLoaded, setIsLoaded] = useState<boolean>(false);
   const [sig, setSig] = useState("");
@@ -38,7 +36,7 @@ const SettingsPage: React.FC = () => {
       .finally(() => {
         setIsLoaded(true);
       });
-  }, []);
+  }, [credential]);
 
   const onSubmit = useCallback((values: User) => {
     fetch(`http://paralibrary.digital/api/users/${values.id}`, {
@@ -81,38 +79,13 @@ const SettingsPage: React.FC = () => {
     []
   );
 
-  const requestConfirm = useContext(ConfirmationContext);
-
-  const deleteAccount = useCallback(() => {
-    requestConfirm(
-      () => {
-        fetch(`http://paralibrary.digital/api/users/${user.id}`, {
-          method: "DELETE",
-          credentials: "include",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }).then((res: Response) => {
-          if (res.ok) {
-            logout();
-          }
-        });
-      },
-      {
-        title: "Delete Account",
-        body:
-          "Are you sure you want to delete you entire account? This action is not reversible.",
-      }
-    );
-  }, [user]);
-
   return (
     <PageLayout
       header={<h1>My Settings</h1>}
       footer={
         <Form>
           <h5>Delete Account</h5>
-          <Form.Text>Confirm name to reveal permanent delete option.</Form.Text>
+          <Form.Text>Type name to reveal permanent delete option.</Form.Text>
           <Form.Group as={Form.Row} controlId="signature">
             <Form.Label column sm={2}>
               Name:
@@ -120,14 +93,14 @@ const SettingsPage: React.FC = () => {
             <Col>
               <Form.Control type="text" name="sig" onChange={handleSigChange} />
             </Col>
-            <Col sm={2}>
-              <Collapse in={isLoaded && sig === userFormik.values.name}>
-                <Button variant="danger" onClick={deleteAccount}>
-                  Delete
-                </Button>
-              </Collapse>
+            <Col sm={4}>
+              <DeleteAccountButton
+                user={user}
+                signature={sig}
+                isLoaded={isLoaded}
+              />
             </Col>
-          </Form.Group>{" "}
+          </Form.Group>
         </Form>
       }
     >
